[mips32][pic32mx] cleanup

+++ os/hal/platforms/MIPS-PIC32MX/mcu/pic32mxxx.h
@@ -174,6 +174,6 @@
 #endif
 
 /* Base addresses for non-standard peripherals */
-#define _EXT_BASE_ADDRESS (&CNCON)
+#define _EXT_BASE_ADDRESS ((unsigned long)&CNCON)
   
 #endif
